requirements.txt
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0
yahooquery==2.3.7
httpx==0.27.0
python-dotenv==1.0.1
